intcode vm supports larger numbers

+++ java/src/main/java/com/grizwold/aoc/Day5.java
@@ -5,7 +5,7 @@ import com.grizwold.aoc.intcode.VM;
 import java.util.concurrent.BlockingQueue;
 
 public class Day5 {
-    public int[] run(String program, BlockingQueue<Integer> in, BlockingQueue<Integer> out) {
+    public long[] run(String program, BlockingQueue<Long> in, BlockingQueue<Long> out) {
         VM vm = new VM(in, out);
         return vm.execute(program);
     }
