+++ java/src/main/java/com/grizwold/aoc/intcode/Opcode_01.java
@@ -9,7 +9,7 @@ class Opcode_01 implements Opcode {
     private String opcode;
 
     @Override
-    public boolean matching(int opcodeDef) {
+    public boolean matching(long opcodeDef) {
         opcode = Utils.toString(opcodeDef);
         paramModes = paramModes(opcode);
         return opcode.matches("...01");
@@ -19,10 +19,10 @@ class Opcode_01 implements Opcode {
     public void execute(VM vm) {
         System.out.printf("@%s {%s} ", vm.instructionPointer, opcode);
 
-        int arg1 = getArg1(vm, paramModes);
-        int arg2 = getArg2(vm, paramModes);
-        int result = arg1 + arg2;
-        int resultPointer = getValue(vm.instructionPointer + 3, IMMEDIATE, vm);
+        long arg1 = getArg1(vm, paramModes);
+        long arg2 = getArg2(vm, paramModes);
+        long result = arg1 + arg2;
+        int resultPointer = (int) getValue(vm.instructionPointer + 3, IMMEDIATE, vm);
 
         vm.memory[resultPointer] = result;
         vm.instructionPointer += 4;
