+++ java/src/main/java/com/grizwold/aoc/intcode/Utils.java
@@ -3,7 +3,7 @@ package com.grizwold.aoc.intcode;
 import org.apache.commons.lang3.StringUtils;
 
 class Utils {
-    static String toString(int opcode) {
+    static String toString(long opcode) {
         String strOpcode = String.valueOf(opcode);
         return StringUtils.leftPad(strOpcode, 5, '0');
     }
